Validate the final output node's type
