+++ src/vellum/workflows/nodes/utils.py
@@ -1,11 +1,14 @@
 from functools import cache
 import json
+import re
 import sys
 from types import ModuleType
-from typing import Any, Callable, Optional, Type, TypeVar, Union, get_args, get_origin
+from typing import Any, Callable, ForwardRef, Optional, Type, TypeVar, Union, get_args, get_origin
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ValidationError
 
+from vellum.workflows.errors.types import WorkflowErrorCode
+from vellum.workflows.exceptions import NodeException
 from vellum.workflows.nodes import BaseNode
 from vellum.workflows.nodes.bases.base_adornment_node import BaseAdornmentNode
 from vellum.workflows.ports.port import Port
@@ -141,3 +144,50 @@ def parse_type_from_str(result_as_str: str, output_type: Any) -> Any:
             raise ValueError("Invalid JSON format for result_as_str")
 
     raise ValueError(f"Unsupported output type: {output_type}")
+
+
+def _get_type_name(obj: Any) -> str:
+    if isinstance(obj, type):
+        return obj.__name__
+
+    if get_origin(obj) is Union:
+        children = [_get_type_name(child) for child in get_args(obj)]
+        return " | ".join(children)
+
+    return str(obj)
+
+
+def cast_to_output_type(result: Any, output_type: Any) -> Any:
+    if isinstance(output_type, ForwardRef) or output_type is Any:
+        # Treat ForwardRefs as Any for now
+        return result
+
+    is_valid_output_type = isinstance(output_type, type)
+    if get_origin(output_type) is Union:
+        allowed_types = get_args(output_type)
+        for allowed_type in allowed_types:
+            try:
+                return cast_to_output_type(result, allowed_type)
+            except NodeException:
+                continue
+    elif get_origin(output_type) is list:
+        allowed_item_type = get_args(output_type)[0]
+        if isinstance(result, list):
+            return [cast_to_output_type(item, allowed_item_type) for item in result]
+    elif is_valid_output_type and issubclass(output_type, BaseModel) and not isinstance(result, output_type):
+        try:
+            return output_type.model_validate(result)
+        except ValidationError as e:
+            raise NodeException(
+                code=WorkflowErrorCode.INVALID_OUTPUTS,
+                message=re.sub(r"\s+For further information visit [^\s]+", "", str(e)),
+            ) from e
+    elif is_valid_output_type and isinstance(result, output_type):
+        return result
+
+    output_type_name = _get_type_name(output_type)
+    result_type_name = _get_type_name(type(result))
+    raise NodeException(
+        code=WorkflowErrorCode.INVALID_OUTPUTS,
+        message=f"Expected an output of type '{output_type_name}', but received '{result_type_name}'",
+    )
